Refactor openid -> adgansplatformen
And put it where the other clients reside. No reason to do it differntly
than the others...
