--- src/core/utils/helpers/openId/OpenIdServiceHttpError.ts
@@ -1,9 +1,0 @@
-import FetcherCriticalHttpError from "../../../fetchers/FetcherCriticalHttpError";
-
-// This HttpError is critical because data such as availability of materials
-// and managing user data is crucial for the functioning of the application.
-export default class OpenIdServiceHttpError<
-  ContextType
-> extends FetcherCriticalHttpError<ContextType> {
-  public readonly name = "OpenIdServiceHttpError";
-}
